+++ docs/user/dql/functions.rst
@@ -2712,13 +2712,13 @@ Return type: INTEGER
 
 Example::
 
-    os> SELECT YEARWEEK('2020-08-26')
+    os> SELECT YEARWEEK('2020-08-26'), YEARWEEK('2019-01-05', 0),
     fetched rows / total rows = 1/1
-    +--------------------------+
-    | YEARWEEK('2020-08-26')   |
-    |--------------------------|
-    | 202034                   |
-    +--------------------------+
+    +--------------------------+-----------------------------+
+    | YEARWEEK('2020-08-26')   | YEARWEEK('2019-01-05', 0)   |
+    |--------------------------+-----------------------------+
+    | 202034                   | 201852                      |
+    +--------------------------+-----------------------------+
 
 String Functions
 ================
